# Micro optimizations for serialization