--- a/src/libraries/System.Text.Encodings.Web/src/System/Text/Encodings/Web/DefaultJavaScriptEncoderBasicLatin.cs
+++ b/src/libraries/System.Text.Encodings.Web/src/System/Text/Encodings/Web/DefaultJavaScriptEncoderBasicLatin.cs
@@ -99,18 +99,11 @@ namespace System.Text.Encodings.Web
 #endif
             Debug.Assert(textLength > 0 && ptr < end);
 
-            // For performance on certain platforms, avoid referencing the static table for every value.
-            ReadOnlySpan<byte> allowList = AllowList;
             do
             {
                 Debug.Assert(text <= ptr && ptr < (text + textLength));
 
-                char value = *(char*)ptr;
-
-                // NeedsEscaping() is manually inlined below for perf; verify semantics remain consistent.
-                Debug.Assert((value > LastAsciiCharacter || allowList[value] == 0) == NeedsEscaping(value));
-
-                if (value > LastAsciiCharacter || allowList[value] == 0)
+                if (NeedsEscaping(*(char*)ptr))
                 {
                     goto Return;
                 }
@@ -242,6 +235,7 @@ namespace System.Text.Encodings.Web
                 byte* end = ptr + textLength;
 
 #if NETCOREAPP
+
                 if (Sse2.IsSupported && textLength >= Vector128<sbyte>.Count)
                 {
                     goto Vectorized;
@@ -251,16 +245,11 @@ namespace System.Text.Encodings.Web
 #endif
                 Debug.Assert(textLength > 0 && ptr < end);
 
-                // For performance on certain platforms, avoid referencing the static table for every value.
-                ReadOnlySpan<byte> allowList = AllowList;
                 do
                 {
                     Debug.Assert(pValue <= ptr && ptr < (pValue + utf8Text.Length));
 
-                    // NeedsEscaping() is manually inlined below for perf; verify semantics remain consistent.
-                    Debug.Assert((allowList[*ptr] == 0) == NeedsEscaping(*ptr));
-
-                    if (allowList[*ptr] == 0)
+                    if (NeedsEscaping(*ptr))
                     {
                         goto Return;
                     }