Primarily this helps with Blazor performance on the Mono interpreter during serialization.

The writer tests are ~1.4x faster and serialization (which uses the writer) are ~1.25x faster. This is an existing benchmark we have been using that represents a large object graph. These timings taken from Edge browser:
```
Before:
Serialization took 1383 ms	Deserialization took 4074 ms	Writer took 1146 ms

After:
Serialization took 1117 ms	Deserialization took 4055 ms	Writer took 804 ms
```

The main perf benefit is due to avoiding accessing the `static ReadOnlySpan<byte>` table when checking to see if we need to escaper each characters. Instead, a local stack variable is made once. This performance may be able to be fixed in the interpreter @BrzVlad; not sure why this is much slower on Mono interpreter, although it could be taking a lock to ensure the static table is properly initialized.

The non-Mono (core) numbers are a bit faster as well (about 2-3%) due to micro optimizations in the writer and serializer code and was not affected by the escaping change mentioned above. Here's a generic writer test:
```
Before:

|          Method | Formatted | SkipValidation | DataSize |     Mean |   Error |  StdDev |   Median |      Min |      Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |---------- |--------------- |--------- |---------:|--------:|--------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
|  WriteBasicUtf8 |     False |          False |       10 | 642.9 ns | 4.94 ns | 4.62 ns | 641.7 ns | 636.7 ns | 650.4 ns | 0.0179 |     - |     - |     120 B |
| WriteBasicUtf16 |     False |          False |       10 | 705.0 ns | 5.50 ns | 5.15 ns | 705.6 ns | 697.5 ns | 715.5 ns | 0.0167 |     - |     - |     120 B |

After:
|          Method | Formatted | SkipValidation | DataSize |     Mean |   Error |  StdDev |   Median |      Min |      Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |---------- |--------------- |--------- |---------:|--------:|--------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
|  WriteBasicUtf8 |     False |          False |       10 | 630.2 ns | 5.96 ns | 5.58 ns | 631.0 ns | 616.8 ns | 636.9 ns | 0.0174 |     - |     - |     120 B |
| WriteBasicUtf16 |     False |          False |       10 | 692.4 ns | 7.38 ns | 6.90 ns | 693.7 ns | 682.4 ns | 705.0 ns | 0.0168 |     - |     - |     120 B |
```