@BrzVlad 2020-07-15
While this looks good, I would prefer to wait for https://github.com/dotnet/runtime/pull/39195, which should get in today. That change should fix most of the overhead with accessing the AllowedList property. I would expect the inlining of NeedsEscaping as that part of this PR to still have some additional benefit, but it's hard to be sure without testing it.