@steveharter 2020-07-15
> I would prefer to wait for #39195, which should get in today

I was hoping to get this in before the Preview 8 snap today. However it makes sense to wait and see if there is still a gain by caching the reference to the static `ReadOnlySpan` instead of retroactively pulling it out.

> I would expect the inlining of NeedsEscaping as that part of this PR to still have some additional benefit, but it's hard to be sure without testing it.

This also has the benefit of helping with the core case as well (2-3% for all of the other optimizations) so it makes sense to keep that portion.