@BrzVlad 2020-07-15
@steveharter The interpreter PR was just merged. I think we could consider the more controversial inlining of NeedsEscape separately in another PR and have the rest of this merged in time for the snap.